Update GCC to version 8.2.0
This is the most modern version of GCC. It also brings new C++ versions.

patch-gcc_cp_cfns.h is not needed anymore, because the new GCC version
uses C++, which removes the sole purpose of that file. I've also
disposed of old GCC patches.

I've added ISL to GCC build, because it seems to provide some
optimizations.