Fused the prepare scripts into one script.
Also made it a little more verbose. A skeleton was added for future
support of more package managers on Linux.

The system does not use sudo (generally, such scripts should not use
sudo), but issues a warning if sudo is not used where it should. The
warning message is customized for the system.

Removed dependency for automake 1.9 - it is no longer in Debian, which
means it's high time the package using it was updated.

The script is now safe to be run with a shell other than bash, like dash.
The shebang on the top was changed to #!/bin/sh.